requirements.txt
python-dotenv==1.0.0
streamlit==1.18.1
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
fastapi==0.110.2
PyPDF2==3.0.1
langchain_community==0.2.12
langchain==0.2.13
#openai==0.27.6
openai==1.42.0